Changed PROP_SYS_PHYSMEM to 64-bit integer. Use 0750 as CONFIG_DIR_MODE in all cases (thus no dice for others). Moved settings_getpagesize() as compat_pagesize() to lib/misc.[ch]. Added compat_max_fd(). Removed misused sysconf(), RTFM. Added compat_mkdir(). settings_init(): 	Show the memory limit too. 	Show the memory sizes in human-friendly form.

+++ src/if/gnet_property_priv.h
@@ -220,7 +220,7 @@ extern const guint32  search_handle_ignored_files;
 extern const guint32  configured_peermode;
 extern const guint32  current_peermode;
 extern const guint32  sys_nofile;
-extern const guint32  sys_physmem;
+extern const guint64  sys_physmem;
 extern const guint32  dl_queue_count;
 extern const guint32  dl_running_count;
 extern const guint32  dl_active_count;
